requirements.txt
torch>=2.0.0
numpy>=1.21.0
hydra-core>=1.3.0
omegaconf>=2.3.0
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
jupyter>=1.0.0
matplotlib>=3.5.0
wandb>=0.15.0
tqdm>=4.65.0
ipdb
loguru
einops>=0.6.0